requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-CORS
Flask-SocketIO
PyMySQL
python-dotenv
google-generativeai
